HIVE-3149: Refactor azure client
To facilitate more thorough unit testing, refactor the azure client as
follows:
- `GetVMCapabilities()` and `GetVirtualMachineSku()` are *removed* from
  the `Client` interface. This prevents them being mocked, which will
  allow us to unit test their contents. These functions ultimately land
  on `ListResourceSKUs()`, a 2-line wrapper around the only actual SDK
  call in these paths.
- Those two funcs are *moved* to be (lowercase/unexported) methods on
  AzureActuator.
- The test helper `mockGetVMCapabilities()` is refactored to mock out
  the lower-level `ListResourceSKUs()`, which inherently extends
  coverage up the stack into both of the (newly-moved) methods.

A subsequent commit will further expand test coverage of the moved
methods.
